Use Rails Secrets For Warehouse.
Let's lean back on the rails.secrets file for sensible naming of items.
These names are not anywhere near as secret as API keys so an
unencrypted file is a great place for them. Define a us_warehouse_name
key with a default that falls back to the known US Warehouse name.